Step 2 solution added
