+++ Calculator.Wolfieeex/Calculator.Wolfieeex/Calculator.Wolfieeex.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.35004.147
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Calculator.Wolfieeex", "Calculator.Wolfieeex\Calculator.Wolfieeex.csproj", "{BF47179E-011B-4AEF-A5C4-98B4D6FD18F9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Calculator.Wolfieeex", "Calculator.Wolfieeex\Calculator.Wolfieeex.csproj", "{BF47179E-011B-4AEF-A5C4-98B4D6FD18F9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CalculatorLibrary", "CalculatorLibrary\CalculatorLibrary.csproj", "{86CB0E38-9BA5-4A6F-9CCB-1FD7B1D0C6FD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{BF47179E-011B-4AEF-A5C4-98B4D6FD18F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BF47179E-011B-4AEF-A5C4-98B4D6FD18F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BF47179E-011B-4AEF-A5C4-98B4D6FD18F9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{86CB0E38-9BA5-4A6F-9CCB-1FD7B1D0C6FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{86CB0E38-9BA5-4A6F-9CCB-1FD7B1D0C6FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{86CB0E38-9BA5-4A6F-9CCB-1FD7B1D0C6FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{86CB0E38-9BA5-4A6F-9CCB-1FD7B1D0C6FD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
